Fix issue with semver fixed version
