+++ lts.json
@@ -1,2 +1,3 @@
 {
+  "ember-prop-types": "^3.0.0"
 }
